+++ ui/footer/build_arrow_right_element.go
@@ -16,13 +16,33 @@ func buildArrowRightElement(
 	element tview.Primitive,
 	size int,
 ) {
-	arrowRightText := fmt.Sprintf("%c", symbols.ArrowRight)
-	arrowRightStyle := tcell.Style{}.
+	return buildArrowElement(symbols.ArrowRight, background, foreground)
+}
+
+func buildArrowLeftElement(
+	background,
+	foreground tcell.Color,
+) (
+	element tview.Primitive,
+	size int,
+) {
+	return buildArrowElement(symbols.ArrowLeft, background, foreground)
+}
+
+func buildArrowElement(
+	symbol rune,
+	background,
+	foreground tcell.Color,
+) (
+	element tview.Primitive,
+	size int,
+) {
+	arrowText := fmt.Sprintf("%c", symbol)
+	arrowStyle := tcell.Style{}.
 		Background(background).
 		Foreground(foreground)
-	arrowRightView := tview.NewTextView().
-		SetText(arrowRightText).
-		SetTextStyle(arrowRightStyle)
-	marginRight := 2
-	return arrowRightView, marginRight
+	arrowView := tview.NewTextView().
+		SetText(arrowText).
+		SetTextStyle(arrowStyle)
+	return arrowView, 1
 }