+++ ui/producer/parameters.go
@@ -9,6 +9,7 @@ import (
 type Parameters struct {
 	Theme                 theme.Theme
 	OnEscapeCallback      OnEscapeCallbackType
+	OnProgressCallback    func(current, total int64)
 	OnRunFinishedCallback func(error)
 	App                   *tview.Application
 	RunnerConfig          runner.Config