+++ ui/theme/default.go
@@ -71,7 +71,8 @@ func NewDefault() Theme {
 		Footer: FooterTheme{
 			Background:      color.Color("#504945").ToTerminal(),
 			Foreground:      color.Color("#bdae93").ToTerminal(),
-			SuiteForeground: color.Color("#bdae93").ToTerminal(),
+			SuiteBackground: color.Color("#fb4934").ToTerminal(),
+			SuiteForeground: color.Color("#282828").ToTerminal(),
 			RootDirectoryPath: RootDirectoryPathTheme{
 				Background:      color.Color("#bdae93").ToTerminal(),
 				Foreground:      color.Color("#504945").ToTerminal(),